+++ argilla/src/extralit/server/context/vectordb.py
@@ -7,7 +7,7 @@ from weaviate import WeaviateClient
 from weaviate.exceptions import WeaviateStartUpError
 
 
-def get_weaviate_client(http_port=None, http_secure=None, grpc_port=None, grpc_secure=None) -> Optional[WeaviateClient]:
+def get_weaviate_client(http_port=8080, http_secure=None, grpc_port=50051, grpc_secure=None) -> Optional[WeaviateClient]:
     if 'WCS_HTTP_URL' not in os.environ:
         return None
 
@@ -24,14 +24,11 @@ def get_weaviate_client(http_port=None, http_secure=None, grpc_port=None, grpc_s
         grpc_port = grpc_port or grpc_url.port
         grpc_secure = grpc_secure or grpc_url.scheme == 'https'
 
-        if not http_url.hostname or not grpc_url.hostname:
-            return None
-
         weaviate_client = weaviate.connect_to_custom(
-            http_host=http_url.hostname,
+            http_host=http_url.hostname or http_url.path,
             http_port=http_port,
             http_secure=http_secure,
-            grpc_host=grpc_url.hostname,
+            grpc_host=grpc_url.hostname or grpc_url.path,
             grpc_port=grpc_port,
             grpc_secure=grpc_secure,
             auth_credentials=weaviate.auth.AuthApiKey(api_keys[0]) \